% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacterization.R
\docType{package}
\name{CohortCharacterization-package}
\alias{CohortCharacterization}
\alias{CohortCharacterization-package}
\title{CohortCharacterization: Generating aggregate statistics for cohorts}
\description{
An R package for aggregate statistics for cohorts using data in the Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/anthonysena/CohortCharacterization}
  \item Report bugs at \url{https://github.com/anthonysena/CohortCharacterization/issues}
}

}
\author{
\strong{Maintainer}: Anthony Sena \email{sena@ohdsi.org}

Authors:
\itemize{
  \item Chris Knoll
}

}
\keyword{internal}
